requirements.txt
asgiref==3.2.10
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==1.0.3
chardet==4.0.0
cloudinary==1.25.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.1
django-allauth==0.44.0
django-channels==0.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
djangorestframework==3.12.4
Faker==0.7.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
numerize==0.12
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-magic==0.4.22
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.4
whitenoise==5.2.0
zope.interface==5.4.0
